requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
astroid==2.5.3
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.2
discord==1.0.1
discord.py==1.7.1
Flask==1.1.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==3.10.1
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.6.0
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.1.0
nltk==3.6.1
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.7.4
python-decouple==3.4
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.4
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
toml==0.10.2
tqdm==4.60.0
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
